+++ tests/openassetio-python/_openassetio_test.cpp
@@ -5,5 +5,9 @@
 namespace py = pybind11;
 
 void registerPyRetainingSharedPtrTestTypes(const py::module_&);
+void registerHostApiTestTypes(py::module_&);
 
-PYBIND11_MODULE(_openassetio_test, mod) { registerPyRetainingSharedPtrTestTypes(mod); }
+PYBIND11_MODULE(_openassetio_test, mod) {
+  registerPyRetainingSharedPtrTestTypes(mod);
+  registerHostApiTestTypes(mod);
+}