+++ RELEASE_NOTES.md
@@ -151,6 +151,12 @@ v1.0.0-alpha.X
 
 ### Improvements
 
+- Added `openassetio-python` C++-to-Python bridge library, providing
+  a `createPythonPluginSystemManagerImplementationFactory` function,
+  which allows a C++ host to load managers using the Python plugin
+  system.
+  [508](https://github.com/OpenAssetIO/OpenAssetIO/issues/508)
+
 - Made `LogggerInterface::log` non-const.
 
 - Added `MangerFactory` as a new, simpler, mechanism for querying for